+++ nautobot_ssot/jobs/examples.py
@@ -1,16 +1,19 @@
 """Sample data-source and data-target Jobs."""
-from typing import Optional
+from typing import Optional, Mapping
 from uuid import UUID
-
+from django.contrib.contenttypes.models import ContentType
 from django.templatetags.static import static
 from django.urls import reverse
 
 from nautobot.dcim.models import Region, Site
+from nautobot.ipam.models import Prefix
+from nautobot.tenancy.models import Tenant
 from nautobot.extras.jobs import Job, StringVar
 from nautobot.extras.models import Status
 
 from diffsync import DiffSync, DiffSyncModel
 from diffsync.enum import DiffSyncFlags
+
 import requests
 
 from nautobot_ssot.jobs.base import DataMapping, DataSource, DataTarget
@@ -53,6 +56,25 @@ class SiteModel(DiffSyncModel):
     slug: str
     status_slug: str
     region_name: Optional[str]  # may be None
+    description: str
+
+    # Not in _attributes or _identifiers, hence not included in diff calculations
+    pk: Optional[UUID]
+
+
+class PrefixModel(DiffSyncModel):
+    """Shared data model representing a Prefix in either of the local or remote Nautobot instances."""
+
+    # Metadata about this model
+    _modelname = "prefix"
+    _identifiers = ("prefix", "tenant_slug")
+    # To keep this example simple, we don't include **all** attributes of a Prefix here. But you could!
+    _attributes = ("description", "status_slug")
+
+    # Data type declarations for all identifiers and attributes
+    prefix: str
+    tenant_slug: str
+    status_slug: str
     description: str
 
     # Not in _attributes or _identifiers, hence not included in diff calculations
@@ -158,6 +180,49 @@ class SiteRemoteModel(SiteModel):
         return super().delete()
 
 
+class PrefixRemoteModel(PrefixModel):
+    """Implementation of Prefix create/update/delete methods for updating remote Nautobot data."""
+
+    @classmethod
+    def create(cls, diffsync, ids, attrs):
+        """Create a new Prefix in remote Nautobot.
+
+        Args:
+            diffsync (NautobotRemote): DiffSync adapter owning this Prefix
+            ids (dict): Initial values for this model's _identifiers
+            attrs (dict): Initial values for this model's _attributes
+        """
+        diffsync.post(
+            "/api/ipam/prefixes/",
+            {
+                "prefix": ids["prefix"],
+                "tenant": {"slug": ids["tenant_slug"]} if ids["tenant_slug"] else None,
+                "description": attrs["description"],
+                "status": attrs["status_slug"],
+            },
+        )
+        return super().create(diffsync, ids=ids, attrs=attrs)
+
+    def update(self, attrs):
+        """Update an existing Site record in remote Nautobot.
+
+        Args:
+            attrs (dict): Updated values for this record's _attributes
+        """
+        data = {}
+        if "description" in attrs:
+            data["description"] = attrs["description"]
+        if "status_slug" in attrs:
+            data["status"] = attrs["status_slug"]
+        self.diffsync.patch(f"/api/dcim/sites/{self.pk}/", data)
+        return super().update(attrs)
+
+    def delete(self):
+        """Delete an existing Site record from remote Nautobot."""
+        self.diffsync.delete(f"/api/dcim/sites/{self.pk}/")
+        return super().delete()
+
+
 class RegionLocalModel(RegionModel):
     """Implementation of Region create/update/delete methods for updating local Nautobot data."""
 
@@ -252,6 +317,62 @@ class SiteLocalModel(SiteModel):
         return super().delete()
 
 
+class PrefixLocalModel(PrefixModel):
+    """Implementation of Prefix create/update/delete methods for updating local Nautobot data."""
+
+    @classmethod
+    def create(cls, diffsync, ids, attrs):
+        """Create a new Prefix record in local Nautobot.
+
+        Args:
+            diffsync (NautobotLocal): DiffSync adapter owning this Prefix
+            ids (dict): Initial values for this model's _identifiers
+            attrs (dict): Initial values for this model's _attributes
+        """
+        prefix = Prefix(prefix=ids["prefix"], description=attrs["description"])
+        if ids["tenant_slug"]:
+            tenant_obj, _ = Tenant.objects.get_or_create(
+                slug=ids["tenant_slug"],
+                defaults={"name": ids["tenant_slug"]},
+            )
+            prefix.tenant = tenant_obj
+
+        status_obj, _ = Status.objects.get_or_create(
+            slug=attrs["status_slug"],
+            defaults={"name": attrs["status_slug"]},
+        )
+        status_obj.content_types.add(ContentType.objects.get_for_model(Prefix))
+        prefix.status = status_obj
+        prefix.validated_save()
+        return super().create(diffsync, ids=ids, attrs=attrs)
+
+    def update(self, attrs):
+        """Update an existing Prefix record in local Nautobot.
+
+        Args:
+            attrs (dict): Updated values for any of this model's _attributes
+        """
+        prefix = Prefix.objects.get(prefix=self.prefix, tenant__slug=self.tenant_slug)
+        for attr_name in ("description",):
+            if attr_name in attrs:
+                setattr(prefix, attr_name, attrs[attr_name])
+
+        if "status_slug" in attrs:
+            status_obj, _ = Status.objects.get_or_create(
+                defaults={"name": attrs["status_slug"]},
+            )
+            status_obj.content_types.add(ContentType.objects.get_for_model(Prefix))
+            prefix.status = status_obj
+        prefix.validated_save()
+        return super().update(attrs)
+
+    def delete(self):
+        """Delete an existing Prefix record from local Nautobot."""
+        prefix = Prefix.objects.get(prefix=self.prefix, tenant__slug=self.tenant_slug)
+        prefix.delete()
+        return super().delete()
+
+
 # In a more complex Job, you would probably want to move each DiffSync subclass into a separate Python module.
 
 
@@ -265,9 +386,10 @@ class NautobotRemote(DiffSync):
     # Model classes used by this adapter class
     region = RegionRemoteModel
     site = SiteRemoteModel
+    prefix = PrefixRemoteModel
 
     # Top-level class labels, i.e. those classes that are handled directly rather than as children of other models
-    top_level = ("region", "site")
+    top_level = ("region", "site", "prefix")
 
     def __init__(self, *args, url=None, token=None, job=None, **kwargs):
         """Instantiate this class, but do not load data immediately from the remote system.
@@ -288,13 +410,18 @@ class NautobotRemote(DiffSync):
             "Authorization": f"Token {self.token}",
         }
 
+    def _get_api_data(self, url_path: str) -> Mapping:
+        """Returns data from a url_path using pagination."""
+        data = requests.get(f"{self.url}/{url_path}", headers=self.headers, params={"limit": 0}).json()
+        result_data = data["results"]
+        while data["next"]:
+            data = requests.get(data["next"], headers=self.headers, params={"limit": 0}).json()
+            result_data.extend(data["results"])
+        return result_data
+
     def load(self):
         """Load Region and Site data from the remote Nautobot instance."""
-        # To keep the example simple, we disable REST API pagination of results.
-        # In a real job, especially when expecting a large number of results,
-        # we'd want to handle pagination, or use something like PyNautobot.
-        region_data = requests.get(f"{self.url}/api/dcim/regions/", headers=self.headers, params={"limit": 0}).json()
-        for region_entry in region_data["results"]:
+        for region_entry in self._get_api_data("api/dcim/regions/"):
             region = self.region(
                 name=region_entry["name"],
                 slug=region_entry["slug"],
@@ -305,18 +432,28 @@ class NautobotRemote(DiffSync):
             self.add(region)
             self.job.log_debug(message=f"Loaded {region} from remote Nautobot instance")
 
-        site_data = requests.get(f"{self.url}/api/dcim/sites/", headers=self.headers, params={"limit": 0}).json()
-        for site_entry in site_data["results"]:
+        for site_entry in self._get_api_data("api/dcim/sites/"):
             site = self.site(
                 name=site_entry["name"],
                 slug=site_entry["slug"],
-                status_slug=site_entry["status"]["value"] if site_entry["status"] else "active",
+                status_slug=site_entry["status"]["value"],
                 region_name=site_entry["region"]["name"] if site_entry["region"] else None,
                 description=site_entry["description"],
                 pk=site_entry["id"],
             )
             self.add(site)
             self.job.log_debug(message=f"Loaded {site} from remote Nautobot instance")
+
+        for prefix_entry in self._get_api_data("api/ipam/prefixes/"):
+            prefix = self.prefix(
+                prefix=prefix_entry["prefix"],
+                description=prefix_entry["description"],
+                status_slug=prefix_entry["status"]["value"],
+                tenant_slug=prefix_entry["tenant"]["slug"] if prefix_entry["tenant"] else "",
+                pk=prefix_entry["id"],
+            )
+            self.add(prefix)
+            self.job.log_debug(message=f"Loaded {prefix} from remote Nautobot instance")
 
     def post(self, path, data):
         """Send an appropriately constructed HTTP POST request."""
@@ -343,9 +480,10 @@ class NautobotLocal(DiffSync):
     # Model classes used by this adapter class
     region = RegionLocalModel
     site = SiteLocalModel
+    prefix = PrefixLocalModel
 
     # Top-level class labels, i.e. those classes that are handled directly rather than as children of other models
-    top_level = ("region", "site")
+    top_level = ("region", "site", "prefix")
 
     def __init__(self, *args, job=None, **kwargs):
         """Instantiate this class, but do not load data immediately from the local system."""
@@ -369,13 +507,24 @@ class NautobotLocal(DiffSync):
             site_model = self.site(
                 name=site.name,
                 slug=site.slug,
-                status_slug=site.status.slug if site.status else "active",
+                status_slug=site.status.slug,
                 region_name=site.region.name if site.region else None,
                 description=site.description,
                 pk=site.pk,
             )
             self.add(site_model)
             self.job.log_debug(message=f"Loaded {site_model} from local Nautobot instance")
+
+        for prefix in Prefix.objects.all():
+            prefix_model = self.prefix(
+                prefix=str(prefix.prefix),
+                description=prefix.description,
+                status_slug=prefix.status.slug,
+                tenant_slug=prefix.tenant.slug if prefix.tenant else "",
+                pk=prefix.pk,
+            )
+            self.add(prefix_model)
+            self.job.log_debug(message=f"Loaded {prefix_model} from local Nautobot instance")
 
 
 # The actual Data Source and Data Target Jobs are relatively simple to implement
@@ -389,6 +538,11 @@ class ExampleDataSource(DataSource, Job):
         description="Remote Nautobot instance to load Sites and Regions from", default="https://demo.nautobot.com"
     )
     source_token = StringVar(description="REST API authentication token for remote Nautobot instance", default="a" * 40)
+
+    def __init__(self):
+        """Initialize ExampleDataSource."""
+        super().__init__()
+        self.diffsync_flags = self.diffsync_flags | DiffSyncFlags.SKIP_UNMATCHED_DST
 
     class Meta:
         """Metaclass attributes of ExampleDataSource."""
@@ -404,30 +558,18 @@ class ExampleDataSource(DataSource, Job):
         return (
             DataMapping("Region (remote)", None, "Region (local)", reverse("dcim:region_list")),
             DataMapping("Site (remote)", None, "Site (local)", reverse("dcim:site_list")),
+            DataMapping("Prefix (remote)", None, "Prefix (local)", reverse("ipam:prefix_list")),
         )
 
-    def sync_data(self):
-        """Perform data sync into Nautobot."""
-        self.log_info(message="Loading current data from remote Nautobot...")
-        remote_nautobot = NautobotRemote(url=self.kwargs["source_url"], token=self.kwargs["source_token"], job=self)
-        remote_nautobot.load()
+    def load_source_adapter(self):
+        """Method to instantiate and load the SOURCE adapter into `self.source_adapter`."""
+        self.source_adapter = NautobotRemote(url=self.kwargs["source_url"], token=self.kwargs["source_token"], job=self)
+        self.source_adapter.load()
 
-        self.log_info(message="Loading current data from local Nautobot...")
-        local_nautobot = NautobotLocal(job=self)
-        local_nautobot.load()
-
-        # In a more complete example you might make these flags configurable rather than hard-coded
-        diffsync_flags = DiffSyncFlags.CONTINUE_ON_FAILURE | DiffSyncFlags.LOG_UNCHANGED_RECORDS
-
-        self.log_info(message="Calculating diffs...")
-        diff = local_nautobot.diff_from(remote_nautobot, flags=diffsync_flags)
-        self.sync.diff = diff.dict()
-        self.sync.save()
-
-        if not self.kwargs["dry_run"]:
-            self.log_info(message="Syncing from remote Nautobot to local Nautobot...")
-            local_nautobot.sync_from(remote_nautobot, flags=diffsync_flags)
-            self.log_info(message="Sync complete")
+    def load_target_adapter(self):
+        """Method to instantiate and load the TARGET adapter into `self.target_adapter`."""
+        self.target_adapter = NautobotLocal(job=self)
+        self.target_adapter.load()
 
     def lookup_object(self, model_name, unique_id):
         """Look up a Nautobot object based on the DiffSync model name and unique ID."""
@@ -441,6 +583,13 @@ class ExampleDataSource(DataSource, Job):
                 return Site.objects.get(name=unique_id)
             except Site.DoesNotExist:
                 pass
+        elif model_name == "prefix":
+            try:
+                return Prefix.objects.get(
+                    prefix=unique_id.split("__")[0], tenant__slug=unique_id.split("__")[1] or None
+                )
+            except Prefix.DoesNotExist:
+                pass
         return None
 
 
@@ -449,6 +598,11 @@ class ExampleDataTarget(DataTarget, Job):
 
     target_url = StringVar(description="Remote Nautobot instance to update", default="https://demo.nautobot.com")
     target_token = StringVar(description="REST API authentication token for remote Nautobot instance", default="a" * 40)
+
+    def __init__(self):
+        """Initialize ExampleDataTarget."""
+        super().__init__()
+        self.diffsync_flags = self.diffsync_flags | DiffSyncFlags.SKIP_UNMATCHED_DST
 
     class Meta:
         """Metaclass attributes of ExampleDataTarget."""
@@ -462,35 +616,20 @@ class ExampleDataTarget(DataTarget, Job):
     def data_mappings(cls):
         """This Job maps Region and Site objects from the local system to the remote system."""
         return (
-            DataMapping("Region (local)", reverse("dcim:site_list"), "Region (remote)", None),
+            DataMapping("Region (local)", reverse("dcim:region_list"), "Region (remote)", None),
             DataMapping("Site (local)", reverse("dcim:site_list"), "Site (remote)", None),
+            DataMapping("Prefix (local)", reverse("ipam:prefix_list"), "Prefix (remote)", None),
         )
 
-    def sync_data(self):
-        """Perform data sync from Nautobot."""
-        self.log_info(message="Loading current data from local Nautobot...")
-        local_nautobot = NautobotLocal(job=self)
-        local_nautobot.load()
+    def load_source_adapter(self):
+        """Method to instantiate and load the SOURCE adapter into `self.source_adapter`."""
+        self.source_adapter = NautobotLocal(job=self)
+        self.source_adapter.load()
 
-        self.log_info(message="Loading current data from remote Nautobot...")
-        remote_nautobot = NautobotRemote(url=self.kwargs["target_url"], token=self.kwargs["target_token"], job=self)
-        remote_nautobot.load()
-
-        # In a more complete example you might make these flags configurable rather than hard-coded
-        diffsync_flags = (
-            DiffSyncFlags.CONTINUE_ON_FAILURE | DiffSyncFlags.LOG_UNCHANGED_RECORDS | DiffSyncFlags.SKIP_UNMATCHED_DST
-        )
-
-        self.log_info(message="Calculating diffs...")
-        diff = local_nautobot.diff_to(remote_nautobot, flags=diffsync_flags)
-        self.sync.diff = diff.dict()
-        self.sync.save()
-
-        if not self.kwargs["dry_run"]:
-            self.log_info(message="Syncing from local Nautobot to remote Nautobot...")
-            local_nautobot.sync_to(remote_nautobot, flags=diffsync_flags)
-            self.log_info(message="Sync complete")
-        return (None, None)
+    def load_target_adapter(self):
+        """Method to instantiate and load the TARGET adapter into `self.target_adapter`."""
+        self.target_adapter = NautobotRemote(url=self.kwargs["target_url"], token=self.kwargs["target_token"], job=self)
+        self.target_adapter.load()
 
     def lookup_object(self, model_name, unique_id):
         """Look up a Nautobot object based on the DiffSync model name and unique ID."""
@@ -503,5 +642,12 @@ class ExampleDataTarget(DataTarget, Job):
             try:
                 return Site.objects.get(name=unique_id)
             except Site.DoesNotExist:
+                pass
+        elif model_name == "prefix":
+            try:
+                return Prefix.objects.get(
+                    prefix=unique_id.split("__")[0], tenant__slug=unique_id.split("__")[1] or None
+                )
+            except Prefix.DoesNotExist:
                 pass
         return None