+++ nautobot_ssot/models.py
@@ -47,6 +47,18 @@ class Sync(BaseModel):
 
     start_time = models.DateTimeField(blank=True, null=True)
     # end_time is represented by the job_result.completed field
+    source_load_time = models.DurationField(blank=True, null=True)
+    target_load_time = models.DurationField(blank=True, null=True)
+    diff_time = models.DurationField(blank=True, null=True)
+    sync_time = models.DurationField(blank=True, null=True)
+    source_load_memory_final = models.PositiveBigIntegerField(blank=True, null=True)
+    source_load_memory_peak = models.PositiveBigIntegerField(blank=True, null=True)
+    target_load_memory_final = models.PositiveBigIntegerField(blank=True, null=True)
+    target_load_memory_peak = models.PositiveBigIntegerField(blank=True, null=True)
+    diff_memory_final = models.PositiveBigIntegerField(blank=True, null=True)
+    diff_memory_peak = models.PositiveBigIntegerField(blank=True, null=True)
+    sync_memory_final = models.PositiveBigIntegerField(blank=True, null=True)
+    sync_memory_peak = models.PositiveBigIntegerField(blank=True, null=True)
 
     dry_run = models.BooleanField(
         default=False, help_text="Report what data would be synced but do not make any changes"